requirements.txt
fastapi
uvicorn[standard]
python-dotenv
dhanhq
SQLAlchemy
mysql-connector-python>=8.0.26
opencensus-ext-azure>=1.1.9  # For Azure Application Insights
gunicorn
Python-jose
python-multipart
passlib
bcrypt
mlflow
pydantic
apache-airflow
apache-airflow-providers-microsoft-azure
apache-airflow-providers-mysql
apache-airflow-providers-apache-spark
python-artificial-intelligence
requests
pandas
numpy
scikit-learn
joblib
pytest
pytest-asyncio
httpx
alembic
pytest-cov
coverage
pytest-mock
pytest-xdist
black
isort
flake8
mypy
pre-commit
bandit
safety
darglint
tox
boto3
azure-identity
azure-storage-blob
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-containerinstance
azure-mgmt-web
azure-mgmt-sql
azure-mgmt-authorization
azure-mgmt-keyvault
azure-keyvault-secrets
azure-keyvault-keys
azure-core
azure-ai-ml
azure-ai-formrecognizer
azure-ai-textanalytics
azure-cosmos
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
azure-servicebus
azure-functions
azure-functions-worker
distributed
fastapi.middleware.cors
starlette.middleware.cors
python-jose[cryptography]

